refactor(match): split mobile match module [GHI#101]
