+++ src/ui/mobile/modules/match/screens/MatchScreen.tsx
@@ -1,55 +1,24 @@
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 
 import { router } from "expo-router";
-import {
-  Circle,
-  Crown,
-  Flag,
-  Frown,
-  HeartCrack,
-  Home,
-  RotateCcw,
-  Shuffle,
-  TimerOff,
-  Trophy,
-  WifiOff,
-  X,
-} from "lucide-react-native";
-import { AppState, Modal, Pressable, View } from "react-native";
+import { AppState, View } from "react-native";
 
 import { abandonGameUseCase } from "@/application/games/abandonGameUseCase";
 import { findOrCreateGameUseCase } from "@/application/games/findOrCreateGameUseCase";
 import { heartbeatUseCase } from "@/application/games/heartbeatUseCase";
 import { placeMarkUseCase } from "@/application/games/placeMarkUseCase";
 import { timeoutTurnUseCase } from "@/application/games/timeoutTurnUseCase";
-import { resolveAvatarSrc, type UserAvatar } from "@/domain/entities/Avatar";
-import {
-  DEFAULT_GRID_SIZE,
-  type GameId,
-  type MatchState,
-} from "@/domain/entities/Game";
+import type { GameId } from "@/domain/entities/Game";
 import { gameRepository } from "@/infrastructure/convex/repository/gameRepository";
 import { useMobileHeader } from "@/ui/mobile/application/providers/MobileHeaderProvider";
 import { FullPageLoader } from "@/ui/mobile/components/FullPageLoader";
-import { H3, H6, Muted, Small } from "@/ui/mobile/components/Typography";
-import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-} from "@/ui/mobile/components/ui/avatar";
-import { Button } from "@/ui/mobile/components/ui/button";
-import { Card, CardContent } from "@/ui/mobile/components/ui/card";
-import { Icon } from "@/ui/mobile/components/ui/icon";
-import { Text } from "@/ui/mobile/components/ui/text";
-import { cn } from "@/ui/mobile/lib/utils";
+import { MatchActions } from "@/ui/mobile/modules/match/components/MatchActions";
+import { MatchBoard } from "@/ui/mobile/modules/match/components/MatchBoard";
+import { MatchPlayers } from "@/ui/mobile/modules/match/components/MatchPlayers";
+import { MatchResultOverlay } from "@/ui/mobile/modules/match/components/match-result/MatchResultOverlay";
+import { MatchErrorScreen } from "@/ui/mobile/modules/match/screens/MatchErrorScreen";
 import { useGame, useTurnTimer } from "@/ui/shared/match/hooks/useGame";
-import {
-  buildMatchPlayers,
-  getMatchResultViewModel,
-  getOpponentId,
-  resolveCurrentSlot,
-  toDisplayBoard,
-} from "@/ui/shared/match/utils";
+import { getOpponentId, resolveCurrentSlot } from "@/ui/shared/match/utils";
 import { useCurrentUser, useUserById } from "@/ui/shared/user/hooks/useUser";
 import { resolvePlayerLabel } from "@/ui/shared/user/resolvePlayerLabel";
 
@@ -102,33 +71,9 @@ export const MatchScreen = ({ gameId }: MatchScreenProps) => {
     };
   }, [setHeader]);
 
-  const resultViewModel = useMemo(() => {
-    if (!game || !currentUser || !opponentUser) {
-      return null;
-    }
-
-    return getMatchResultViewModel({
-      status: game.status,
-      endedReason: game.endedReason,
-      winner: game.winner,
-      abandonedBy: game.abandonedBy,
-      p1UserId: game.p1UserId,
-      currentUserId,
-      score: game.match?.score,
-      currentUser: {
-        name: resolvePlayerLabel(currentUser, "You"),
-        avatar: currentUser.avatar,
-      },
-      opponentUser: {
-        name: resolvePlayerLabel(opponentUser, "Opponent"),
-        avatar: opponentUser.avatar,
-      },
-    });
-  }, [currentUser, currentUserId, game, opponentUser]);
-
   if (!gameId) {
     return (
-      <CenteredState
+      <MatchErrorScreen
         title="Missing match"
         description="A game id is required before the mobile match screen can load."
       />
@@ -202,25 +147,20 @@ export const MatchScreen = ({ gameId }: MatchScreenProps) => {
     }
   };
 
-  const players = buildMatchPlayers(
-    game.p1UserId,
-    game.currentTurn,
-    currentUser,
-    opponentUser,
-    game.match,
-  );
-
   return (
     <>
       <View className="flex-1 justify-around gap-6 pb-10">
-        <MatchPlayersSection
-          players={players}
-          showWins={game.match.format !== "single"}
+        <MatchPlayers
+          p1UserId={game.p1UserId}
+          currentTurn={game.currentTurn}
           timerActive={Boolean(timerActive)}
           timerProgress={timerProgress}
+          currentUser={currentUser}
+          opponentUser={opponentUser}
+          match={game.match}
         />
 
-        <MatchBoardSection
+        <MatchBoard
           board={game.board}
           gridSize={game.gridSize}
           status={game.status}
@@ -232,7 +172,7 @@ export const MatchScreen = ({ gameId }: MatchScreenProps) => {
           onCellPress={handleCellPress}
         />
 
-        <MatchActionsSection
+        <MatchActions
           match={game.match}
           isAbandoning={isAbandoning}
           onAbandonMatch={handleAbandonMatch}
@@ -240,588 +180,24 @@ export const MatchScreen = ({ gameId }: MatchScreenProps) => {
       </View>
 
       <MatchResultOverlay
-        model={resultViewModel}
+        status={game.status}
+        endedReason={game.endedReason}
+        winner={game.winner}
+        abandonedBy={game.abandonedBy}
+        p1UserId={game.p1UserId}
+        currentUserId={currentUserId}
+        score={game.match?.score}
+        currentUser={{
+          name: resolvePlayerLabel(currentUser, "You"),
+          avatar: currentUser.avatar,
+        }}
+        opponentUser={{
+          name: resolvePlayerLabel(opponentUser, "Opponent"),
+          avatar: opponentUser.avatar,
+        }}
         onPrimaryAction={handleCreateNewGame}
       />
     </>
-  );
-};
-
-type MatchPlayersSectionProps = {
-  players: ReturnType<typeof buildMatchPlayers>;
-  showWins: boolean;
-  timerActive: boolean;
-  timerProgress: number;
-};
-
-const MatchPlayersSection = ({
-  players,
-  showWins,
-  timerActive,
-  timerProgress,
-}: MatchPlayersSectionProps) => {
-  return (
-    <View className="flex-row gap-3">
-      {players.map((player) => (
-        <MatchPlayerCard
-          key={player.id}
-          name={player.name}
-          symbol={player.symbol}
-          wins={player.wins}
-          showWins={showWins}
-          isTurn={player.isTurn}
-          avatar={player.avatar}
-          accent={player.accent}
-          turnTimer={
-            timerActive && player.isTurn
-              ? { isActive: true, progress: timerProgress }
-              : undefined
-          }
-        />
-      ))}
-    </View>
-  );
-};
-
-type MatchPlayerCardProps = {
-  name: string;
-  symbol: "X" | "O";
-  wins: number;
-  showWins: boolean;
-  isTurn: boolean;
-  avatar?: string;
-  accent: "primary" | "opponent";
-  turnTimer?: {
-    isActive: boolean;
-    progress: number;
-  };
-};
-
-const MatchPlayerCard = ({
-  name,
-  symbol,
-  wins,
-  showWins,
-  isTurn,
-  avatar,
-  accent,
-  turnTimer,
-}: MatchPlayerCardProps) => {
-  const showTurnTimer = Boolean(turnTimer?.isActive);
-  const accentClasses =
-    accent === "opponent"
-      ? {
-          badge: "bg-opponent text-opponent-foreground",
-          border: "border-opponent/50",
-          symbol: "text-opponent",
-          timer: "bg-opponent",
-          shadow: "shadow-[0_0_10px_-8px_rgba(84,126,234,0.28)]",
-        }
-      : {
-          badge: "bg-primary text-primary-foreground",
-          border: "border-primary/50",
-          symbol: "text-primary",
-          timer: "bg-primary",
-          shadow: "shadow-[0_0_10px_-8px_rgba(61,168,105,0.28)]",
-        };
-  const timerProgress = Math.max(0, Math.min(turnTimer?.progress ?? 0, 1));
-  const isUrgent = showTurnTimer && timerProgress <= 0.1;
-  const badgeClassName = isUrgent
-    ? "bg-destructive text-destructive-foreground"
-    : accentClasses.badge;
-
-  return (
-    <View className="relative flex-1 pt-2.5">
-      {isTurn ? (
-        <View className="absolute left-1/2 top-0 z-20 -translate-x-1/2">
-          <Small
-            className={cn(
-              "rounded-full px-2.5 py-1 text-[10px] uppercase tracking-[0.18em]",
-              badgeClassName,
-            )}
-          >
-            Turn
-          </Small>
-        </View>
-      ) : null}
-      <Card
-        className={cn(
-          "flex-1 gap-0 rounded-3xl border-border/60 px-0 py-0 shadow-none",
-          {
-            "opacity-70": !isTurn,
-            [accentClasses.border]: isTurn,
-            [accentClasses.shadow]: isTurn,
-          },
-        )}
-      >
-        {showTurnTimer ? (
-          <View className="overflow-hidden rounded-t-3xl">
-            <View className="h-1 w-full bg-secondary/80">
-              <View
-                className={cn("h-full", {
-                  "bg-destructive": isUrgent,
-                  [accentClasses.timer]: !isUrgent,
-                })}
-                style={{ width: `${Math.max(timerProgress * 100, 0)}%` }}
-              />
-            </View>
-          </View>
-        ) : null}
-        <CardContent className="items-center gap-3.5 px-4 pb-4 pt-5">
-          <Avatar
-            alt={name}
-            className={cn("size-20 border-2 bg-card", {
-              [accentClasses.border]: isTurn,
-              "border-border/60": !isTurn,
-            })}
-          >
-            {avatar ? <AvatarImage source={{ uri: avatar }} /> : null}
-            <AvatarFallback className="bg-secondary">
-              <H6>{name.slice(0, 1)}</H6>
-            </AvatarFallback>
-          </Avatar>
-          <View className="items-center gap-1">
-            <View className="flex-row items-center gap-2">
-              <H6 className="text-center text-sm">{name}</H6>
-              <Text
-                className={cn("text-sm font-semibold", accentClasses.symbol)}
-              >
-                ({symbol})
-              </Text>
-            </View>
-            {showWins ? (
-              <Muted className="mt-0 text-xs text-muted-foreground">
-                Wins: {wins}
-              </Muted>
-            ) : null}
-          </View>
-        </CardContent>
-      </Card>
-    </View>
-  );
-};
-
-type MatchBoardSectionProps = {
-  board: Array<"P1" | "P2" | null>;
-  gridSize?: number;
-  status: "waiting" | "playing" | "paused" | "ended" | "canceled";
-  currentTurn: "P1" | "P2";
-  currentUserId?: string;
-  p1UserId: string;
-  isTimeUp: boolean;
-  isPlacing: boolean;
-  onCellPress: (index: number) => void;
-};
-
-const MatchBoardSection = ({
-  board,
-  gridSize,
-  status,
-  currentTurn,
-  currentUserId,
-  p1UserId,
-  isTimeUp,
-  isPlacing,
-  onCellPress,
-}: MatchBoardSectionProps) => {
-  const resolvedGridSize = gridSize ?? DEFAULT_GRID_SIZE;
-  const currentSlot = resolveCurrentSlot(currentUserId, p1UserId);
-  const isCurrentUserP1 = currentSlot ? currentSlot === "P1" : true;
-  const displayBoard = toDisplayBoard(board, resolvedGridSize);
-  const markSize = resolvedGridSize <= 4 ? 76 : 56;
-  const currentPlayerClasses = {
-    P1: isCurrentUserP1 ? "text-primary" : "text-opponent",
-    P2: isCurrentUserP1 ? "text-opponent" : "text-primary",
-  };
-
-  return (
-    <Card
-      className={cn(
-        "relative rounded-[2rem] border-border/70 px-4 py-4 shadow-none",
-        {
-          "border-destructive/60": isTimeUp,
-        },
-      )}
-      style={{
-        shadowColor: "#000000",
-        shadowOffset: { width: 0, height: 12 },
-        shadowOpacity: 0.12,
-        shadowRadius: 8,
-        elevation: 2,
-      }}
-    >
-      <View className="gap-2">
-        {displayBoard.map((row, rowIndex) => (
-          <View key={`row-${rowIndex.toString()}`} className="flex-row gap-2">
-            {row.map((cell, colIndex) => {
-              const isCellTaken = cell !== "";
-              const isKnownUser = Boolean(currentSlot);
-              const isNotYourTurn =
-                status === "playing" &&
-                isKnownUser &&
-                currentTurn !== currentSlot &&
-                !isTimeUp;
-              const isHardDisabled =
-                status !== "playing" || !isKnownUser || isPlacing || isTimeUp;
-              const isDisabled =
-                isHardDisabled || isCellTaken || Boolean(isNotYourTurn);
-              const index = rowIndex * resolvedGridSize + colIndex;
-
-              return (
-                <Pressable
-                  key={`${rowIndex.toString()}-${colIndex.toString()}`}
-                  accessibilityRole="button"
-                  accessibilityLabel={`Match cell ${index + 1}`}
-                  disabled={isHardDisabled}
-                  className={cn(
-                    "aspect-square flex-1 items-center justify-center rounded-3xl border border-border/70 bg-secondary",
-                    {
-                      "opacity-55": isDisabled,
-                    },
-                  )}
-                  onPress={() => {
-                    if (isCellTaken || isNotYourTurn) {
-                      return;
-                    }
-
-                    onCellPress(index);
-                  }}
-                >
-                  <BoardCellMark
-                    mark={cell}
-                    size={markSize}
-                    className={
-                      cell === "X"
-                        ? currentPlayerClasses.P1
-                        : currentPlayerClasses.P2
-                    }
-                  />
-                </Pressable>
-              );
-            })}
-          </View>
-        ))}
-      </View>
-      {isTimeUp ? (
-        <View className="absolute inset-0 items-center justify-center rounded-[2rem] bg-card/92 px-6">
-          <View className="items-center gap-4">
-            <View className="size-18 items-center justify-center rounded-full border border-destructive/30 bg-destructive/10">
-              <Icon as={TimerOff} className="size-9 text-destructive" />
-            </View>
-            <View className="items-center gap-1">
-              <H3 className="text-center uppercase tracking-[0.2em] text-destructive">
-                Time&apos;s up
-              </H3>
-              <Muted className="mt-0 text-center text-sm">
-                Your turn expired before the move could be placed.
-              </Muted>
-            </View>
-          </View>
-        </View>
-      ) : null}
-    </Card>
-  );
-};
-
-type BoardCellMarkProps = {
-  mark: string;
-  size: number;
-  className: string;
-};
-
-const BoardCellMark = ({ mark, size, className }: BoardCellMarkProps) => {
-  if (!mark) {
-    return null;
-  }
-
-  return (
-    <View
-      className="items-center justify-center"
-      style={{ width: size, height: size }}
-    >
-      <Icon
-        as={mark === "X" ? X : Circle}
-        className={className}
-        size={size}
-        strokeWidth={3}
-      />
-    </View>
-  );
-};
-
-type MatchActionsSectionProps = {
-  match: MatchState | null;
-  isAbandoning: boolean;
-  onAbandonMatch: () => Promise<void>;
-};
-
-const MatchActionsSection = ({
-  match,
-  isAbandoning,
-  onAbandonMatch,
-}: MatchActionsSectionProps) => {
-  const showRoundCounter = Boolean(match && match.format !== "single");
-  const bestOf = match ? match.targetWins * 2 - 1 : 1;
-
-  return (
-    <View className="gap-4">
-      {showRoundCounter ? (
-        <View className="items-center">
-          <View className="rounded-full border border-border/60 bg-card px-4 py-2">
-            <Small className="text-[10px] font-semibold uppercase tracking-[0.18em] text-muted-foreground">
-              Round {match?.roundIndex ?? 1} · Best of {bestOf}
-            </Small>
-          </View>
-        </View>
-      ) : null}
-
-      <Button
-        variant="outline"
-        className="h-12 rounded-full border-destructive/30 bg-secondary"
-        onPress={() => void onAbandonMatch()}
-        disabled={isAbandoning}
-      >
-        <Icon as={Flag} className="text-destructive" size={16} />
-        <Text className="font-semibold text-destructive">
-          {isAbandoning ? "Abandoning..." : "Abandon Match"}
-        </Text>
-      </Button>
-    </View>
-  );
-};
-
-type MatchResultOverlayProps = {
-  model: ReturnType<typeof getMatchResultViewModel>;
-  onPrimaryAction: () => Promise<void>;
-};
-
-const MatchResultOverlay = ({
-  model,
-  onPrimaryAction,
-}: MatchResultOverlayProps) => {
-  if (!model) {
-    return null;
-  }
-
-  const iconConfig = resolveResultIcon(model.icon, model.accent);
-  const currentAccentClassName = model.isCurrentWinner
-    ? "text-primary"
-    : "text-destructive";
-  const opponentAccentClassName = model.isCurrentWinner
-    ? "text-destructive"
-    : "text-primary";
-
-  return (
-    <Modal transparent visible animationType="fade">
-      <View className="flex-1 justify-center bg-background/95 px-6">
-        <Card className="rounded-[2rem] border-border/70 px-0 py-0 shadow-[0_18px_35px_-26px_rgba(0,0,0,0.5)]">
-          <CardContent className="gap-6 px-6 py-6">
-            <View className="items-center gap-2">
-              <View
-                className={cn(
-                  "size-20 items-center justify-center rounded-full",
-                  iconConfig.containerClassName,
-                )}
-              >
-                <Icon
-                  as={iconConfig.icon}
-                  className={iconConfig.iconClassName}
-                />
-              </View>
-              <H3
-                className={cn(
-                  "text-center uppercase tracking-[0.18em]",
-                  model.accent === "primary"
-                    ? "text-primary"
-                    : "text-destructive",
-                )}
-              >
-                {model.title}
-              </H3>
-              {model.subtitle ? (
-                <Muted className="mt-0 text-center whitespace-pre-line text-sm leading-6">
-                  {model.subtitle}
-                </Muted>
-              ) : null}
-              {model.pill ? (
-                <Small className="rounded-full bg-secondary px-3 py-1 text-[10px] uppercase tracking-[0.18em] text-muted-foreground">
-                  {model.pill}
-                </Small>
-              ) : null}
-            </View>
-
-            <View className="rounded-[1.5rem] border border-border/60 bg-secondary/50 px-4 py-4">
-              <View className="flex-row items-center justify-between gap-4">
-                <ResultPlayer
-                  label="You"
-                  name={model.currentUser.name}
-                  avatar={model.currentUser.avatar}
-                  score={model.score?.current}
-                  accentClassName={currentAccentClassName}
-                  isWinner={model.isCurrentWinner}
-                />
-                <View className="items-center gap-1">
-                  <Small className="text-[10px] uppercase tracking-[0.18em] text-muted-foreground">
-                    Final
-                  </Small>
-                  <Small className="text-[10px] uppercase tracking-[0.18em] text-muted-foreground">
-                    Result
-                  </Small>
-                </View>
-                <ResultPlayer
-                  label="Opponent"
-                  name={model.opponentUser.name}
-                  avatar={model.opponentUser.avatar}
-                  score={model.score?.opponent}
-                  accentClassName={opponentAccentClassName}
-                  isWinner={!model.isCurrentWinner}
-                  align="right"
-                />
-              </View>
-              {model.footer ? (
-                <View className="mt-4 border-t border-border/60 pt-4">
-                  <Small className="text-center text-[10px] uppercase tracking-[0.18em] text-primary">
-                    {model.footer}
-                  </Small>
-                </View>
-              ) : null}
-            </View>
-
-            <View className="gap-3">
-              <Button
-                className={cn(
-                  "h-12 rounded-full",
-                  model.accent === "primary" ? "bg-primary" : "bg-destructive",
-                )}
-                onPress={() => void onPrimaryAction()}
-              >
-                <Icon as={RotateCcw} className="text-primary-foreground" />
-                <Text className="font-semibold uppercase tracking-[0.18em] text-primary-foreground">
-                  {model.primaryLabel}
-                </Text>
-              </Button>
-              <Button
-                variant="secondary"
-                className="h-12 rounded-full"
-                onPress={() => router.replace("/play")}
-              >
-                <Icon as={Shuffle} />
-                <Text className="font-semibold uppercase tracking-[0.18em]">
-                  {model.changeModeLabel}
-                </Text>
-              </Button>
-              <Button
-                variant="outline"
-                className="h-12 rounded-full"
-                onPress={() => router.replace("/")}
-              >
-                <Icon as={Home} />
-                <Text className="font-semibold uppercase tracking-[0.18em]">
-                  {model.secondaryLabel}
-                </Text>
-              </Button>
-            </View>
-          </CardContent>
-        </Card>
-      </View>
-    </Modal>
-  );
-};
-
-type ResultPlayerProps = {
-  label: string;
-  name: string;
-  avatar?: UserAvatar;
-  score?: number;
-  accentClassName: string;
-  isWinner: boolean;
-  align?: "left" | "right";
-};
-
-const ResultPlayer = ({
-  label,
-  name,
-  avatar,
-  score,
-  accentClassName,
-  isWinner,
-  align = "left",
-}: ResultPlayerProps) => {
-  const avatarSrc = resolveAvatarSrc(avatar);
-
-  return (
-    <View
-      className={cn("min-w-0 flex-1 gap-2", {
-        "items-end": align === "right",
-        "items-start": align === "left",
-      })}
-    >
-      <Avatar
-        alt={name}
-        className={cn("size-14 border-2 bg-card", {
-          "border-border/60": !isWinner,
-          "border-primary/70": isWinner,
-          "border-destructive/70": !isWinner,
-        })}
-      >
-        {avatarSrc ? <AvatarImage source={{ uri: avatarSrc }} /> : null}
-        <AvatarFallback className="bg-secondary">
-          <H6>{name.slice(0, 1)}</H6>
-        </AvatarFallback>
-      </Avatar>
-      <View
-        className={cn(
-          "absolute top-10 size-5 items-center justify-center rounded-full",
-          {
-            "-right-1 bg-primary": align === "left" && isWinner,
-            "-right-1 bg-destructive": align === "left" && !isWinner,
-            "-left-1 bg-primary": align === "right" && isWinner,
-            "-left-1 bg-destructive": align === "right" && !isWinner,
-          },
-        )}
-      >
-        <Icon
-          as={isWinner ? Crown : Frown}
-          className="size-3 text-primary-foreground"
-        />
-      </View>
-      <View
-        className={cn("gap-1", {
-          "items-end": align === "right",
-          "items-start": align === "left",
-        })}
-      >
-        <Small className="text-[10px] uppercase tracking-[0.18em] text-muted-foreground">
-          {label}
-        </Small>
-        <H6 className="text-sm" numberOfLines={1}>
-          {name}
-        </H6>
-        {typeof score === "number" ? (
-          <Text className={cn("text-xl font-semibold", accentClassName)}>
-            {score}
-          </Text>
-        ) : null}
-      </View>
-    </View>
-  );
-};
-
-type CenteredStateProps = {
-  title: string;
-  description: string;
-};
-
-const CenteredState = ({ title, description }: CenteredStateProps) => {
-  return (
-    <View className="flex-1 justify-center">
-      <Card className="rounded-3xl border-border/60">
-        <CardContent className="gap-2 px-6 py-6">
-          <H3 className="text-left text-xl">{title}</H3>
-          <Muted className="mt-0 text-base leading-6">{description}</Muted>
-        </CardContent>
-      </Card>
-    </View>
   );
 };
 
@@ -933,42 +309,4 @@ const useGameHeartbeat = (gameId?: GameId) => {
       stop();
     };
   }, [gameId]);
-};
-
-const resolveResultIcon = (
-  icon: NonNullable<ReturnType<typeof getMatchResultViewModel>>["icon"],
-  accent: NonNullable<ReturnType<typeof getMatchResultViewModel>>["accent"],
-) => {
-  if (icon === "wifi") {
-    return {
-      icon: WifiOff,
-      iconClassName: "size-10 text-amber-500",
-      containerClassName: "bg-amber-500/12",
-    };
-  }
-
-  if (icon === "flag") {
-    return {
-      icon: Flag,
-      iconClassName: "size-10 text-muted-foreground",
-      containerClassName: "bg-secondary",
-    };
-  }
-
-  if (icon === "trophy") {
-    return {
-      icon: Trophy,
-      iconClassName: "size-10 text-primary",
-      containerClassName:
-        accent === "primary"
-          ? "bg-primary/12 border border-primary/20"
-          : "bg-destructive/12 border border-destructive/20",
-    };
-  }
-
-  return {
-    icon: HeartCrack,
-    iconClassName: "size-10 text-destructive",
-    containerClassName: "bg-destructive/12 border border-destructive/20",
-  };
 };
